[svn-r13064]
Fix several bugs

1) the parsing of subsetting was using atoi to convert the parameter to an int, which caused problems for numbers greater that int. Substitute with atof

2) the printing of indices in the subsetting case was not being done. Solution: calculate the element position at the start of the subsetting using the algorythm 

Given an index I(z,y,x) its position from the beginning of an array   of sizes A(size_z, size_y,size_x) is given by
Position of I(z,y,x) = index_z * size_y * size_x 
                              + index_y * size_x
                              + index_x

And pass that position to the function that dumps data, h5tools_dump_simple_data. 

3) several index counters were declared as int, use hsize_t instead

4) modified the test generation program so that it includes test cases for subsetting of 1d, 2d, 3d, and 4d arrays and add these tests to the shell script

+++ tools/h5dump/h5dump.c
@@ -2962,7 +2962,7 @@ parse_hsize_list(const char *h_list)
     for (ptr = h_list; i < size_count && ptr && *ptr && *ptr != ';' && *ptr != ']'; ptr++)
         if (isdigit(*ptr)) {
             /* we should have an integer now */
-            p_list[i++] = (hsize_t)atoi(ptr);
+            p_list[i++] = (hsize_t)atof(ptr);
 
             while (isdigit(*ptr))
                 /* scroll to end of integer */
